Function trust_dns_rustls::tls_stream::tls_connect

pub fn tls_connect(
    name_server: SocketAddr, 
    dns_name: String, 
    client_config: Arc<ClientConfig>
) -> (Box<dyn Future<Item = TlsStream<ClientSession>, Error = Error> + Send>, BufStreamHandle)

Creates a new TlsStream to the specified name_server

RFC 7858, DNS over TLS, May 2016

3.2.  TLS Handshake and Authentication

  Once the DNS client succeeds in connecting via TCP on the well-known
  port for DNS over TLS, it proceeds with the TLS handshake [RFC5246],
  following the best practices specified in [BCP195].

  The client will then authenticate the server, if required.  This
  document does not propose new ideas for authentication.  Depending on
  the privacy profile in use (Section 4), the DNS client may choose not
  to require authentication of the server, or it may make use of a
  trusted Subject Public Key Info (SPKI) Fingerprint pin set.

  After TLS negotiation completes, the connection will be encrypted and
  is now protected from eavesdropping.

Arguments

name_server - IP and Port for the remote DNS resolver

dns_name - The DNS name, Subject Public Key Info (SPKI) name, as associated to a certificate